requirements.txt
ujson
bcrypt
nose==1.3.4
importlib; python_version < '2.7'
pg8000
sqlalchemy==1.0.12
argparse
ase==3.11
spglib==1.9.4.2
tornado==4.3.0
sockjs-tornado
websocket-client
futures
httplib2
